requirements.txt
Django==4.2.14
DjangoRestFramework==3.14.0